Got rid of the nasty bind mounts.
This means that you no longer need to provision when running vagrant up.
If you had a previously provisioned machine and you run vagrant up, it
will simply use the old machine. If you do want a new machine you can
run either vagrant reload --provision or vagrant up --provision
depending on whether your machine is running or not.

This also means that the configuration files will be in the folders
wwwconfig and settings_override. Its not as clean, but it works much
better than forcing provisioning every time.