Split vagrant into two vms
There is now a vm for the web server and one for mysql. They currently
have 2gb of ram each, but that can be changed independently.

For the most part, this doesn't do much for a default build, but if you
happen to be running with live data, it makes a huge difference.